Make module for class Pinger and resolve linux problem
