+++ main.py
@@ -2,7 +2,7 @@
 # coding: utf-8
 
 from argparse import ArgumentParser
-from platform import system
+from module.pinger import Pinger
 from re import match
 import subprocess
 import sys
@@ -54,61 +54,6 @@ def check_args() -> str:
     else:
         sys.exit("Le paramètre dois être un nom de fichier !")
 
-
-class Pinger:
-    """Une classe pour représenter un Ping.
-
-    Args:
-       system_now (str): système d'exploitation actuel.
-        count (int): nombre de ping (par défaut 4).
-        nbr_second (int): nombre de seconde pour chaque ping (par défaut 4).
-
-    Attributes:
-        __system_now (str): système d'exploitation actuel.
-        __count (int): nombre de ping (par défaut 4).
-        __nbr_second (int): nombre de seconde pour chaque ping (par défaut 4).
-
-    Returns:
-        str: La commande ping avec la version du système d'exploitation actuel.
-    """
-
-    def __init__(self, system_now=system().lower(), count=4, nbr_second=4):
-        self.__system_now = system_now
-
-        if isinstance(count, int):
-            self.__count = count
-        else:
-            sys.exit("Pas un entier !")
-
-        if isinstance(count, int):
-            self.__nbr_second = nbr_second
-        else:
-            sys.exit("Pas un entier !")
-
-    def number_of_ping(self) -> str:
-        """nombre de seconde pour chaque ping.
-            Returns:
-                str: Le nombre de ping demandé pour la version du système d'exploitation actuel.
-        """
-        if self.__system_now == "windows":
-            return f"-n {self.__count}"
-        else:
-            return f"-c {self.__count}"
-
-    def timeout(self) -> str:
-        """timeout pour chaque ping.
-            Returns:
-                str: Le nombre de timeout pour chaque ping avec la version du système d'exploitation actuel.
-        """
-        if self.__system_now == "windows":
-            return f"-w {(self.__nbr_second * 1000)}"
-        else:
-            return f"-w {self.__nbr_second}"
-
-    def __str__(self):
-        return f"ping {self.number_of_ping()} {self.timeout()}"
-
-
 def ping(hostname: str) -> bool:
     """Effectuer la commande ping sur les sites.
 
@@ -126,7 +71,7 @@ def ping(hostname: str) -> bool:
     test1 = Pinger()
     commande = f"{test1} {hostname}"
     try:
-        output = subprocess.check_output(commande, shell=False, universal_newlines=True)
+        output = subprocess.check_output(commande, shell=True, universal_newlines=True)
 
         return "0%" in output
